+++ AnsAuctions/CHANGELOG.md
@@ -1,3 +1,8 @@
+AnS [Auctions] Changes 1.0.2
+-----------------------------
+* Updated auction posting to match the new PostAuction api changes by blizzard.
+* Fixed issue where the time duration was not being retrieved properly before trying to post an auction.
+
 AnS [Auctions] Changes 1.0.1
 -----------------------------
 * Fixed issue when selling items from multiple listed stacks, the list would not update properly and remove the items listed.